Generate generics with macros instead of decoders
This commit replaces the macros which generated circe Encoder and
Decoders for scrooge-generated classes for thrift structs and unions
with new macros which instead generate shapeless LabelledGeneric values.
These values match the values that shapeless would generate
(mostly; there are some changes to match the existing behaviour with
default values) for the equivalent case classes, which allows us to use
circe’s semi-auto derivation to get the encoders and decoders.

This has the advantage that we no longer need to maintain as much macro
complexity, and we can now create encoders/decoders for recursive struct
hierarchies (either direct or indirect).

Hopefully this will also mean we can migrate to scala 3 more easily, as
our macros are less complicated and more connected to shapeless, which
we hear scala 3 macros are inspired by.

Unfortunately, circe’s fully automatic derivation no longer works with
this approach, which means this is a (small, I think) breaking change
for consumers of the library.

Implementation notes:

- To get it working I had to use whitebox macros instead of blackbox
  macros, but unfortunately the existing macros don’t work when they’re
  whitebox, so I’ve split the new macros into their own file (I can’t
  see how to have whitebox and blackbox macros in the same file).

- This method no longer suppresses nulls in the output, but I believe
  that’s possible for consumers using a circe option.

- I suspect there’s a way to replace these macros altogether with
  shapeless-style implicits, but I don’t know how yet and this
  works.

- To match the existing decoder behaviour, any struct field with a
  default parameter is allowed to be missing from the json, in which
  case the default value will be used.