Develop generic attempt
I made a few more things explicit and it’s getting closer (for example,
I can define what looks like the right labelled generic format in a way
the compiler accepts), but I think I’ve hit a wall.

In particular, I can call `deriveReprAsObjectEncoder` for type HNil, but
not for the tagged generic type I’m trying (even when I’ve copied that
from the type of something that the derivation succeeds for, like a
tuple of String and Integer).

I think at this point it’s a complicated macro problem that’s beyond me
for now.