Try delaying evaluation of implicitly
When I print the result of the macro, `implicitly` has already been
resolved, which I think is what’s causing our loop: each macro call
attempts to resolve the implicit, resulting in another macro call before
the current one has finished.

If we could instead delay the evaluation of the next macro call until
the end of the current one (or avoid it entirely in a base case,
somehow?) then the loop might finish.