Use whitebox macro instead of blackbox
This works! I have to instantiate the encoders and decoders
manually (semi-automatically), which is unfortunate, but it looks like
the macro being whitebox is the solution to the generic macro not being
successful.

Unfortunately, this clashes with the existing macros a bit: the
thriftEnum or thriftUnion macros get called for the recursive type and
get stuck in a loop, producing a stack overflow. Possibly providing a
generic representation for those as well will fix everything? I’m a
little worried they’ll all be called, so we might have to merge them
into a single macro that determines whether the argument is a struct,
enum, or union, and proceed as required.

I also don’t know if requiring users to use semi-auto derivation is
going to be feasible, but it is something we considered as an option
initially!