Manually write union generic
In this commit I have successfully written a manual generic instance for
a union type! Plus demonstrated that as long as you have decoders for
the enums and unions in descendant fields, the generic approach for
structs works for StructA.

However, changing the macros to whitebox does appear to break the
existing enum/union macros: this is why I had to manually create the
encoder/decoder in the tests for Enum (based on what the
encodeThriftEnum macro _would_ output). Without commenting out the enum
and union implicits, I get a stack overflow from what appears to be an
infinite loop in the macro – I don’t know why the macros are even being
called, so it’s a bit of a mystery. I’m hoping that replacing both of
them with macros that create generics will avoid/fix this problem.

Another alternative might be to make just the new macro whitebox, and
keep the existing one blackbox: I think I could do this by putting them
in different files?

I’ve also noticed that the struct decoders don’t seem to be accumulating
errors, which is disappointing: this is a regression from the previous
ones.