Document behaviour with list and map types
For list (and map) fields, the thrift-generated apply method has a
default value (of the empty container) even when the thrift field
definition doesn’t have a default value.

This means that our generated decoders allow leaving the field out (with
the same result as providing the empty array or empty object).