Encode and decode UnknownUnionField values
The classes scrooge generates for thrift unions include
`UnknownUnionField` values, which are for writing code that doesn’t
crash when new fields/variants are added.

The existing behaviour in fezziwig for these values is that the encoder
successfully encodes them (as null), but the decoder throws an error
when decoding null (unless the field is also marked optional).

Rather than mimic the existing behaviour (which I also don’t know how to
do), I’ve chosen to preserve the field metadata (names, ids, and types)
but drop the field values. This seems like a sensible way to handle
unknown union fields for now, as preserving the values would require a
bit more effort. (I guess we’d need to convert the raw bytes to/from
circe’s Json type and store that? Hmmm…)