Update tests to avoid DecodingFailure equality
The error messages have changed, so these tests need updating (Maybe the
tests shouldnʼt depend on the error message? Not sure), but also it
seems the definition of equality for DecodingFailure has changed: I
couldnʼt get the failures constructed in the tests to compare equal to
the ones from circe. So Iʼve changed these tests to explicitly compare
the message and history.