Ignore scrooge updates past 21.8.0
We use some thrift reserved words in our thrift definitions, and as of
[scrooge version
21.9.0](https://github.com/twitter/scrooge/releases/tag/scrooge-21.9.0),
scrooge throws errors when it encounters these reserved words.

Until we’ve worked out how to get around this problem, this commit
configures Scala Steward not to update scrooge past the version we’re
currently on (21.8.0).

For more information, see:

- https://github.com/guardian/content-api/issues/2903
- https://github.com/guardian/content-api-models/pull/231